Add files via upload
An information note about admin login (for edit or delete operations) was added to the AdminPanel and to the Warning section of the GuidePanel.

Alternatively, new information titled "Warning" was provided.

The popup for the edit section in the AdminPanel was structured within a ScrollPanel.

An asterisk (*) was added to the end of required fields in the AddArtifact section.

A required format was specified for the Current and Discovery Location fields. Alerts were added for incorrect entries where numeric input is expected.

All code has been reviewed.